refactor: Phase 2 타입 모듈 정규화 1차 적용
