+++ src/core/index.ts
@@ -12,7 +12,8 @@ export { NpmDownloader, getNpmDownloader } from './downloaders/npm';
 
 // Packager
 export { ArchivePackager, getArchivePackager } from './packager/archive-packager';
-export type { ArchiveOptions, ArchiveProgress, PackageManifest } from './packager/archive-packager';
+export type { ArchiveOptions, ArchiveProgress } from './packager/archive-packager';
+export type { PackageManifest } from '../types/manifest/package-manifest';
 
 export { ScriptGenerator, getScriptGenerator } from './packager/script-generator';
 export type { ScriptOptions, GeneratedScript } from './packager/script-generator';
