+++ src/core/packager/archive-packager.ts
@@ -2,8 +2,11 @@ import archiver from 'archiver';
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import { PackageInfo } from '../../types';
+import type { PackageManifest } from '../../types/manifest/package-manifest';
 import logger from '../../utils/logger';
 import { resolvePath, toUnixPath } from '../shared/path-utils';
+
+export type { PackageManifest } from '../../types/manifest/package-manifest';
 
 // 압축 형식 타입
 export type ArchiveFormat = 'zip' | 'tar.gz';
@@ -24,15 +27,6 @@ export interface ArchiveProgress {
   processedBytes: number;
   totalBytes: number;
   percentage: number;
-}
-
-// 패키지 매니페스트
-export interface PackageManifest {
-  version: string;
-  createdAt: string;
-  packages: PackageInfo[];
-  totalSize: number;
-  fileCount: number;
 }
 
 export class ArchivePackager {
