+++ src/types/download/result.ts
@@ -1,0 +1,10 @@
+import type { DownloadItem } from './item';
+
+export interface DownloadResult {
+  success: boolean;
+  items: DownloadItem[];
+  totalDownloaded: number;
+  totalFailed: number;
+  totalSkipped: number;
+  outputPath?: string;
+}
