+++ src/types/manifest/package-manifest.ts
@@ -1,0 +1,9 @@
+import type { PackageInfo } from '../package-manager/metadata';
+
+export interface PackageManifest {
+  version: string;
+  createdAt: string;
+  packages: PackageInfo[];
+  totalSize: number;
+  fileCount: number;
+}
