+++ src/types/package-manager/package-manager.ts
@@ -1,0 +1,9 @@
+export type PackageType =
+  | 'pip'
+  | 'conda'
+  | 'maven'
+  | 'npm'
+  | 'yum'
+  | 'apt'
+  | 'apk'
+  | 'docker';
