+++ src/types/platform/architecture.ts
@@ -1,0 +1,11 @@
+export type Architecture =
+  | 'x86_64'
+  | 'amd64'
+  | 'arm64'
+  | 'aarch64'
+  | 'i386'
+  | 'i686'
+  | 'noarch'
+  | 'all'
+  | 'arm/v7'
+  | '386';
